Fix GnuPG encryption
Since I'm mostly enrypting data for myself, set encrypyt-to-self as default behaviour.

In emacs, set myself as the recipient; if I don't do this, emacs defaults to symmetric encryption,
which is a guaranteed data loss even if I inadvertently use my GnuPG passphrase out of instinct; it
it is more likely than not that I'll mistype and never be able to decrypt my
symmetrically-encrypted file.