fix!: Revert system.stateSystem back to versions on install
BREAKING CHANGE: Apparently I misread or didn't realize that I wasn't supposed to change the
system.stateVersion. I assumed I had to upgrade them to use later versions of
config, but that is apparently not true. I will be recreating most of these
instances anyway, but for now, move back to the original versions when these
systems were created.