feat: cache GnuPG keys for longer.
It is annoying to constantly have to re-input GnuPG PINs every time I
load up emacs or open a new personal file.

Use gnupg-agent's caching options os that if I use a GPG key within an
hour of the last time I used it, I do not need to login, up to a
maxium of 24 hours, thus forcing me to log in at least once a day.