fix: attach ID property to project element.
Because I want each root headline to be a TODO, I want the ID attached to it.

The original code here attached the ID to the top of the file, which I hadn't
populated with a title, so all new projects had an invalid blank name and not
the title of the appropriate headline.